[#100974462] - Removed Footer
I removed the footer on 'View Profile' and 'Edit Profile' because the
links to all the pages in the footer already exist. The only one I had
to add was the a link the 'Cards' page listing the library of all the
cards; the link is now on the top right of the nav bar.